# docs/swagger.yaml
definitions:
  BaseSuccessResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
    type: object
...
  SuccessResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  UserEditRequest:
    properties:
      address:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
...
  UserResponse:
    properties:
      address:
        type: string
      email:
        type: string
      id:
        type: string
      is_admin:
        type: boolean
      name:
        type: string
      username:
        type: string
    type: object
...
  model.Product:
    properties:
      base_price:
        type: integer
      created_at:
        type: integer
      descriptions:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      sell_price:
        type: integer
      stock:
        type: integer
      updated_at:
        type: integer
    type: object
  model.Transaction:
    properties:
      created_at:
        type: integer
      id:
        type: string
      is_active:
        type: boolean
      is_paid:
        type: boolean
      paid_at:
        type: integer
      products:
        items:
          $ref: '#/definitions/model.TransactionProduct'
        type: array
      total_item:
        type: integer
      total_margin:
        type: integer
      total_sum:
        type: integer
      user_id:
        type: string
    type: object
  model.TransactionProduct:
    properties:
      id:
        type: string
      item:
        $ref: '#/definitions/model.Product'
      item_id:
        type: string
      margin:
        type: integer
      quantity:
        type: integer
      sum:
        type: integer
      transaction_id:
        type: string
    type: object
  model.TransactionSummaryResponse:
    properties:
      total_margin_sold_product:
        type: integer
      total_quantity_sold_product:
        type: integer
      total_sum_sold_product:
        type: integer
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
  /admin/transaction/{id}:
    get:
      consumes:
      - application/json
      description: Get transaction detail by id
      parameters:
      - description: token
        in: header
        name: Cookie
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.Transaction'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get Transaction by id
      tags:
      - Transaction
    patch:
      consumes:
      - application/json
      description: Verify transaction detail
      parameters:
      - description: uuid
        in: path
        name: id
        required: true
        type: string
      - description: token
        in: header
        name: Cookie
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/BaseSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Verify Transaction
      tags:
      - Transaction
  /admin/transaction/all:
    get:
      consumes:
      - application/json
      description: Get All transaction detail
      parameters:
      - description: token
        in: header
        name: Cookie
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Transaction'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get  All Transaction
      tags:
      - Transaction
  /admin/transaction/summary:
    get:
      consumes:
      - application/json
      description: Get Summary of All Sold Product
      parameters:
      - description: token
        in: header
        name: Cookie
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.TransactionSummaryResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get Summary Sold Product
      tags:
      - Transaction
...
  /transaction:
    get:
      consumes:
      - application/json
      description: Get New transaction detail
      parameters:
      - description: token
        in: header
        name: Cookie
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.Transaction'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Detail Transaction
      tags:
      - Transaction
    post:
      consumes:
      - application/json
      description: Create New Transaction
      parameters:
      - description: token
        in: header
        name: Cookie
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.Transaction'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Create Transaction
      tags:
      - Transaction
  /transaction/{id}:
    get:
      consumes:
      - application/json
      description: Get By Id Self Own Detail Transaction
      parameters:
      - description: token
        in: header
        name: Cookie
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.Transaction'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get By Id Own Transaction
      tags:
      - Transaction
  /transaction/all:
    get:
      consumes:
      - application/json
      description: Get All Self Own Detail Transaction
      parameters:
      - description: token
        in: header
        name: Cookie
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Transaction'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get All Own Transaction
      tags:
      - Transaction
  /user/:
    delete:
      description: Delete logged In self request user
      parameters:
      - description: token
        in: header
        name: Cookie
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Delete self request user
      tags:
      - User
    get:
      description: get logged in request user
      parameters:
      - description: token
        in: header
        name: Cookie
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: get self request user
      tags:
      - User
    put:
      description: Update self request user data
      parameters:
      - description: User
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/UserEditRequest'
      - description: token
        in: header
        name: Cookie
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Update self user
      tags:
      - User
swagger: "2.0"